refactor progress types and quiz hook
