+++ amplify_outputs.json
@@ -1,0 +1,1 @@
+{}
