+++ src/services/client.ts
@@ -1,4 +1,5 @@
 import { generateClient } from 'aws-amplify/data';
+import type { Client } from 'aws-amplify/data';
 import type { Schema } from '../../amplify/data/resource';
 
-export const client = generateClient<Schema>();
+export const client: Client<Schema> = generateClient<Schema>();
